separates radio controller access types
Every radio controller provides access to the SRAM, TRXFIFO and different
registers. How this access is realized, is defined by the manufacturer and
described in the data sheet of the radio controller. To ease implementation
of a radio controller the different access types (SRAM, FIFO, register)
contained in the specfile are seperated.